LayoutTests/platform/ios-simulator/editing/selection/move-left-right-expected.txt

CONSOLE MESSAGE: line 186: TypeError: null is not an object (evaluating 'positions[i].node.tagName')
abc
אבג

abc

אבג
abcאבגdef
אבגabcדהו
abcאבגדהו
אבגabcdef
١٢٣ابة
ابة١٢٣
abcאבגdef
אבגabcדהו
abcאבג123דהוdef
abcאבג123
abcאבג123def
אבג123דהוabcזחט456יכל
before אחרי אנציקלופדיה
לפני after encyclopedia
before אחרי אנציקלופדיה
לפני after encyclopedia
This is יותר צר מיתר the boxes.
This is יותר צר מיתר the boxes.
Lorem
ipsum
צלחת
מצנפת
abcdefאבגדהו
אבגדהוabcdef
abcאבגדהוghi
אבגabcdefדהו
abcאבגדהו
אבגabcdef
abcאבגdef
אבגabcדהו
abcאבגdef
אבגabcדהו
abcאdef
אבגaדהו
abcאבגdef
אבגabcדהו
abcאבגdef
אבגabcדהו
abc
def
אבג
דהו
abcקקק123נננdef
Test 1, LTR:
  Moving right: "\n        abc\n    "[9, 10, 11, 12]
  Moving left:  "\n        abc\n    "[12, 11, 10, 9]
Test 1, RTL:
  Moving left: "\n        abc\n    "[9, 11, 10, 12]
  Moving right:  "\n        abc\n    "[12, 10, 11, 9]
Test 2, LTR:
  Moving right: 
